refactor: consolidate hooks structure
