+++ src/pages/dashboard/AttendantDashboardPage.tsx
@@ -7,9 +7,9 @@ import { useState } from 'react';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { useStations } from '@/hooks/api/useStations';
-import { useReadings } from '@/hooks/api/useReadings';
-import { useFuelPrices } from '@/hooks/api/useFuelPrices';
+import { useStations } from '@/hooks/useStations';
+import { useReadings } from '@/hooks/useReadings';
+import { useFuelPrices } from '@/hooks/useFuelPrices';
 import { format } from 'date-fns';
 import { Loader2, FileText, Fuel, CreditCard, DollarSign } from 'lucide-react';
 import { Link } from 'react-router-dom';
