+++ src/pages/dashboard/CashReportPage.tsx
@@ -11,9 +11,9 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { useToast } from '@/hooks/use-toast';
-import { useStations } from '@/hooks/api/useStations';
-import { useCreditors } from '@/hooks/api/useCreditors';
-import { useSubmitCashReport } from '@/hooks/api/useAttendant';
+import { useStations } from '@/hooks/useStations';
+import { useCreditors } from '@/hooks/useCreditors';
+import { useSubmitCashReport } from '@/hooks/useAttendant';
 import { format } from 'date-fns';
 import { ArrowLeft, Plus, Trash, DollarSign, CreditCard, Loader2 } from 'lucide-react';
 import { CashReport, CreditEntry } from '@/api/services/attendantService';
