+++ src/pages/dashboard/CashReportsListPage.tsx
@@ -7,8 +7,8 @@ import { useNavigate } from 'react-router-dom';
 import { useRoleGuard } from '@/hooks/useRoleGuard';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
-import { useStations } from '@/hooks/api/useStations';
-import { useCashReports } from '@/hooks/api/useAttendant';
+import { useStations } from '@/hooks/useStations';
+import { useCashReports } from '@/hooks/useAttendant';
 import { format } from 'date-fns';
 import { ArrowLeft, RefreshCw, Loader2, DollarSign } from 'lucide-react';
 import { Badge } from '@/components/ui/badge';
