+++ src/pages/dashboard/DashboardPage.tsx
@@ -8,10 +8,10 @@ import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { RefreshCw, TrendingUp, Users, Fuel, Loader2, Building2, Shield, Package, BarChart3 } from 'lucide-react';
 import { useAuth } from '@/contexts/AuthContext';
-import { useStations } from '@/hooks/api/useStations';
-import { usePumps } from '@/hooks/api/usePumps';
-import { useFuelPrices } from '@/hooks/api/useFuelPrices';
-import { useReadings } from '@/hooks/api/useReadings';
+import { useStations } from '@/hooks/useStations';
+import { usePumps } from '@/hooks/usePumps';
+import { useFuelPrices } from '@/hooks/useFuelPrices';
+import { useReadings } from '@/hooks/useReadings';
 import { useAdminDashboard } from '@/hooks/useDashboard';
 import { useDashboardAnalytics } from '@/hooks/useAnalytics';
 import { useSystemHealth } from '@/hooks/useSystemHealth';
