+++ src/pages/dashboard/FuelInventoryPage.tsx
@@ -8,10 +8,10 @@ import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { Fuel, AlertTriangle, TrendingDown, RefreshCw, Loader2, FileSpreadsheet } from 'lucide-react';
-import { useStations } from '@/hooks/api/useStations';
+import { useStations } from '@/hooks/useStations';
 import { useFuelInventory, useFuelInventorySummary } from '@/hooks/useFuelInventory';
 import { useDeliveriesInventory } from '@/hooks/useFuelDeliveries';
-import { useGenerateReport } from '@/hooks/api/useReports';
+import { useGenerateReport } from '@/hooks/useReports';
 import { Link, useNavigate } from 'react-router-dom';
 import { useToast } from '@/hooks/use-toast';
 import { useRoleGuard } from '@/hooks/useRoleGuard';
