+++ src/pages/dashboard/StationsPage.tsx
@@ -10,7 +10,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Badge } from '@/components/ui/badge';
 import { Plus, Eye, Fuel } from 'lucide-react';
-import { useStations } from '@/hooks/api/useStations';
+import { useStations } from '@/hooks/useStations';
 import { useRoleGuard } from '@/hooks/useRoleGuard';
 
 export default function StationsPage() {
